Use os.walk rather than os.listdir to find dsym folders.
-- Before, we effectively assumed that the dsym folders existed at
the top level of the downloaded archive. However, that isn't always
the case. Some customers have archives containing a single folder, which
in turn contains the dsym folders. In such cases, the upload script
fails to find and upload the symbols, even though they are there.

-- Therefore, we make the script more robust by using os.walk, allowing
us to search the entire symbol archive for dsym folders, rather than
just the top level.

-- Note that because the search_root now no longer necessarily
corresponds to the directory just about each dsym folder, we
change directory to the root of each dsym before adding it to the
tar file for upload. This way, the upload preserves the same
directory structure as before this change.